+++ Source/DataTransfer.Legacy.MassImport.NUnit.Integration/FunctionalTests/ChoicesTests.cs
@@ -12,6 +12,7 @@ using FieldHelper = MassImport.NUnit.Integration.Helpers.FieldHelper;
 using MassImportManager = Relativity.MassImport.Api.MassImportManager;
 using Moq;
 using Relativity.API;
+using Relativity.Productions.Services.Private.V1;
 
 namespace MassImport.NUnit.Integration.FunctionalTests
 {
@@ -62,7 +63,7 @@ namespace MassImport.NUnit.Integration.FunctionalTests
 			var artifactManager = new ArtifactManager();
 			var baseContext = CoreContext.ChicagoContext;
 
-			_sut = new MassImportManager(AssemblySetup.TestLogger, artifactManager, baseContext, new Mock<IHelper>().Object);
+			_sut = new MassImportManager(AssemblySetup.TestLogger, artifactManager, baseContext, HelperMock.Object);
 		}
 
 		[TearDown]