+++ Source/DataTransfer.Legacy.MassImport.NUnit.Integration/FunctionalTests/CollationEdgeCasesTests.cs
@@ -10,6 +10,7 @@ using Relativity.MassImport.Api;
 using MassImportManager = Relativity.MassImport.Api.MassImportManager;
 using Moq;
 using Relativity.API;
+using Relativity.Productions.Services.Private.V1;
 
 namespace MassImport.NUnit.Integration.FunctionalTests
 {
@@ -34,7 +35,7 @@ namespace MassImport.NUnit.Integration.FunctionalTests
 			var artifactManager = new ArtifactManager();
 			var baseContext = CoreContext.ChicagoContext;
 
-			_sut = new MassImportManager(AssemblySetup.TestLogger, artifactManager, baseContext, new Mock<IHelper>().Object);
+			_sut = new MassImportManager(AssemblySetup.TestLogger, artifactManager, baseContext, HelperMock.Object);
 		}
 
 		[Test]